+++ src/Microsoft.OpenApi.Readers/ParseNodes/AnyFieldMapParameter.cs
@@ -1,0 +1,40 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. 
+
+using System;
+using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
+
+namespace Microsoft.OpenApi.Readers.ParseNodes
+{
+    internal class AnyFieldMapParameter<T>
+    {
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        public AnyFieldMapParameter(
+            Func<T, IOpenApiAny> propertyGetter,
+            Func<T, IOpenApiAny, IOpenApiAny> propertySetter,
+            Func<T, OpenApiSchema> schemaGetter)
+        {
+            this.PropertyGetter = propertyGetter;
+            this.PropertySetter = propertySetter;
+            this.SchemaGetter = schemaGetter;
+        }
+
+        /// <summary>
+        /// Function to retrieve the value of the property.
+        /// </summary>
+        public Func<T, IOpenApiAny> PropertyGetter { get; }
+
+        /// <summary>
+        /// Function to set the value of the property.
+        /// </summary>
+        public Func<T, IOpenApiAny, IOpenApiAny> PropertySetter { get; }
+
+        /// <summary>
+        /// Function to get the schema to apply to the property.
+        /// </summary>
+        public Func<T, OpenApiSchema> SchemaGetter { get; }
+    }
+}
