+++ src/Microsoft.OpenApi.Readers/ParseNodes/AnyListFieldMapParameter.cs
@@ -1,0 +1,41 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. 
+
+using System;
+using System.Collections.Generic;
+using Microsoft.OpenApi.Any;
+using Microsoft.OpenApi.Models;
+
+namespace Microsoft.OpenApi.Readers.ParseNodes
+{
+    internal class AnyListFieldMapParameter<T>
+    {
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        public AnyListFieldMapParameter(
+            Func<T, IList<IOpenApiAny>> propertyGetter,
+            Func<T, IList<IOpenApiAny>, IList<IOpenApiAny>> propertySetter,
+            Func<T, OpenApiSchema> schemaGetter)
+        {
+            this.PropertyGetter = propertyGetter;
+            this.PropertySetter = propertySetter;
+            this.SchemaGetter = schemaGetter;
+        }
+
+        /// <summary>
+        /// Function to retrieve the value of the property.
+        /// </summary>
+        public Func<T, IList<IOpenApiAny>> PropertyGetter { get; }
+
+        /// <summary>
+        /// Function to set the value of the property.
+        /// </summary>
+        public Func<T, IList<IOpenApiAny>, IList<IOpenApiAny>> PropertySetter { get; }
+
+        /// <summary>
+        /// Function to get the schema to apply to the property.
+        /// </summary>
+        public Func<T, OpenApiSchema> SchemaGetter { get; }
+    }
+}
